+++ apps/web/src/components/transaction-history.tsx
@@ -1,7 +1,7 @@
 import { useClassName } from '../hooks/use-class-name';
 
 import { Spinner } from './spinner';
-import { ArrowDownCircle, ArrowUpCircle } from 'lucide-react';
+import { ArrowDown, ArrowDownCircle, ArrowUp, ArrowUpCircle } from 'lucide-react';
 
 import { useNavigate } from 'react-router-dom';
 import { useAccount, useTransactions } from '@cbdc-markka/utils-react';
@@ -53,7 +53,12 @@ export function TransactionHistory() {
       {isPending ? (
         <Spinner />
       ) : account?.transactions.length > 0 ? (
-        generateTransactionList()
+        account?.transactions.map((t, i) => (
+          <Transaction
+            data={t}
+            key={i}
+          />
+        ))
       ) : (
         <div className='flex flex-col justify-center items-center flex-1'>
           <h2 className='text-sm text-slate-500'>Ei tapahtumia.</h2>
@@ -67,34 +72,38 @@ function Transaction({ data }: { data: TTransaction }) {
   const { account } = useAccount();
   const navigate = useNavigate();
   const received = data.to === account?.id;
-  const amt = (data.amount_in_cents / 100).toFixed(2);
+  const amt = (data.amount_in_cents / 100) * (received ? 1 : -1);
 
   const amountClassName = useClassName(
-    received ? 'text-green-400' : 'text-red-400',
-    'font-mono flex gap-2 items-center',
+    received ? 'text-green-400' : 'text-slate-600',
+    'font-mono flex gap-2 items-center text-sm',
   );
-  const amtString = received ? `+` : `-`;
+
   return (
     <div
-      className='bg-white py-2 px-4 flex w-full gap-4 items-center cursor-pointer border-b border-slate-200'
-      onClick={() => navigate('transaction/' + data.id)}>
-      {received ? <ArrowDownCircle color='green' /> : <ArrowUpCircle color='red' />}
-      <div className='flex flex-col'>
-        <div className={amountClassName}>
-          <span>{amtString}</span>
-          <div className='flex items-baseline'>
-            <CurrencySymbol size={10} />
-            <span>
-              {' '}
-              {Number(amt).toLocaleString('fi', {
-                minimumFractionDigits: 2,
-                maximumFractionDigits: 2,
-              })}
-            </span>
-          </div>
+      className='bg-white py-4 px-4 flex w-full gap-4 items-center cursor-pointer border-b border-slate-200 justify-between'
+      onClick={() => navigate('/auth/transaction/' + data.id)}>
+      <div className='flex items-center gap-4'>
+        {received ? <ArrowDown color='green' /> : <ArrowUp color='red' />}
+        <div className='flex flex-col'>
+          <span className='text-xs'>{new Date(data.timestamp).toLocaleDateString('fi')}</span>
+          <span className='text-xs text-slate-500'>
+            {received ? data.from_email : data.to_email}
+          </span>
         </div>
-
-        <span className='text-xs text-slate-500'>{received ? data.from_email : data.to_email}</span>
+      </div>
+      <div className={amountClassName}>
+        <div className='flex items-baseline'>
+          <span>
+            {' '}
+            {Number(amt).toLocaleString('fi', {
+              minimumFractionDigits: 2,
+              maximumFractionDigits: 2,
+              signDisplay: 'always',
+            })}
+          </span>
+          <span>mk</span>
+        </div>
       </div>
     </div>
   );