+++ apps/web/src/screen/overview-screen.tsx
@@ -1,0 +1,60 @@
+import { useAccount } from '@cbdc-markka/utils-react';
+import { CurrencySymbol } from '../components/currency';
+import { Spinner } from '../components/spinner';
+import { TransactionHistory } from '../components/transaction-history';
+import { useAnimatedNumber } from '../hooks/use-animated-number';
+import { useNavigate } from 'react-router-dom';
+import { Button } from '../components/button';
+import { LogOut } from 'lucide-react';
+import { AppScreen } from '../components/app-screen';
+
+export function OverviewScreen() {
+  const navigate = useNavigate();
+
+  return (
+    <AppScreen headerShown={false}>
+      <main className='flex flex-col w-full bg-linear-to-b from-primary to-indigo-500 flex-1 grow-0 max-h-full h-full'>
+        <div className='w-full flex py-12 px-4 items-center justify-between'>
+          <WalletContainer />
+          <div className='flex items-center gap-4'>
+            <Button
+              compact
+              circular
+              type='button'
+              variant='ghost'
+              onClick={() => navigate('/auth/logout')}>
+              <LogOut color='white' />
+            </Button>
+          </div>
+        </div>
+        <div className='w-full flex-1 flex-col flex bg-white overflow-y-scroll rounded-t-2xl'>
+          <TransactionHistory />
+        </div>
+      </main>
+    </AppScreen>
+  );
+}
+
+function WalletContainer() {
+  const { account, isPending } = useAccount();
+  const currentBalance = useAnimatedNumber(!isPending ? account.balance_in_cents / 100 : 0);
+  return (
+    <div className='flex flex-col'>
+      <h3 className='text-white text-sm'>Tilin Saldo</h3>
+      <h2 className='text-white text-3xl font-mono flex gap-2 items-baseline'>
+        <CurrencySymbol
+          strokeWidth={0.9}
+          size='1.2rem'
+        />
+        {isPending ? (
+          <Spinner />
+        ) : (
+          Number(currentBalance).toLocaleString('fi', {
+            minimumFractionDigits: 2,
+            maximumFractionDigits: 2,
+          })
+        )}
+      </h2>
+    </div>
+  );
+}